float.

+++ gamutrf/sigfinder.py
@@ -389,7 +389,7 @@ def argument_parser():
                         help='UDP stream port')
     parser.add_argument('--bin_mhz', default=20, type=int,
                         help='monitoring bin width in MHz')
-    parser.add_argument('--max_raw_power', default=50, type=int,
+    parser.add_argument('--max_raw_power', default=50, type=float,
                         help='maximum raw power permitted from FFT')
     parser.add_argument('--width', default=10, type=int,
                         help=f'minimum signal width to detect a peak (multiple of {SCAN_FRES / 1e6} MHz, e.g. 10 is {10 * SCAN_FRES / 1e6} MHz)')
